Improve contrast between failed and incomplete color
Commit c3af6a39f2e72559c045666d1bafd749001de538 increased the contrast
of progress bar segments showing the number of failures by making the
color we generally use for failures darker. However, that decreased the
difference between the color of failures and incompletes which makes
them harder to distinguish. So this commit reverts this original change.
To still have a good contrast in the progress bar segment the text color
within the progress bar is set to black. This looks good in bright and
dark mode (the mode actually makes no difference for the progress bars).

Related ticket: https://progress.opensuse.org/issues/128783