Fix sporadic failures of comments test
This will hopefully fix https://progress.opensuse.org/issues/128153
which still sometimes occurs. It explicitly waits until the text form is
no longer empty. Maybe this is not sufficient but for now I'll try to
keep this workaround as simple as possible. (It is a workaround because
`send_keys` should actually only return after the text is present.)