Only output JSON when using --json-output
If --json-output is used, users expect only JSON and no additional text,
so the JSON can easily be parsed.
It was printed to stderr, so it can be treated separately, but still,
such text output should only appear when for example in a verbose mode.

Issue: https://progress.opensuse.org/issues/128405