Add indexed column `webhook_id` to scheduled products table
This is required to be able to associate a scheduled product
with an entity from a webhook payload, e.g. a PR on GitHub.
That is in turn required to be able to find a possibly already
existing scheduled product when a webhook is called, e.g. to
implement cancellation and cleanup in certain cases.

Related ticket: https://progress.opensuse.org/issues/127949