Consider a tag's version when cleaning up job results
When cleaning up job results, jobs being part of an important
build are treated differently. So far really only the BUILD
variable is used to decide whether a job belongs to an important
build - even if the according tag includes a version number.

With this change the version from the tag is considered as well
and matched against the job's VERSION variable. When a tag only
specifies the BUILD then it is still just matched against the
job's BUILD variable.

Related ticket/note: https://progress.opensuse.org/issues/129244#note-18