Fix openqa-clone-custom-git-refspec on git-checkout job sources
If openqa-clone-custom-git-refspec is called with openQA jobs that already
specify a git repository in CASEDIR the product dir is computed by the
openQA worker as a relative path without the git checkout directory as a
prefix. That so far has led to the problem that the product directory
string that is passed to openqa-clone-job
would be concatenated without separator leading to cases like
"my_distriproducts/foo" when instead "my_distri" and "products/foo"
should be joint with "/" as standard path separator. This commit fixes
this within openqa-clone-custom-git-refspec by always joining the path
components with "/" as path separator while ensuring that no slashes are
duplicated themselves.

Additionally tested by manually scheduling openQA jobs
openqa.opensuse.org with the generated product dir compontents.

Related progress issue: https://progress.opensuse.org/issues/130096