Fix race condition at the end of developer mode test
We must wait for the job result as the job state is immediately
set to cancelled once the cancellation is triggered. So at this
point logs might not have been uploaded yet. Only if the result
has been set by the worker logs are supposed to be present and
the check for that can succeed.