contrib: Add autoyast profile for SUSE PRG2 o3 workers
In the new SUSE PRG2 datacenter we have new servers with 2x512GiB NVMe
and 1x6TiB NVMe. I think it is not feasible to combine all those with
the existing autoyast profile template so adding as separate file. The
profile also explicitly adds the "kvm_server" pattern so that a network
bridge is configured already from the start.

By the way I tried out to use the .xml.erb file supplied to the openSUSE
installer which could not parse the file. So maybe the .xml.erb profile
is meant to be processed by a server into a final xml file or a
different command than the boot line `autoyast=…` would need to be used.

Related progress issue: https://progress.opensuse.org/issues/132134