Remove max_conns setting again from Nginx config
Unfortunately `max_conns` does not make much sense for us without
the `queue` setting, which is only available in the commercial version
of Nginx currently. In the open source version any new incoming request
would immediately receive a 502 response if all 30 upstream connections
are in use.

However if we ever run into a problem where the openqa-webui service
gets hopelessly overloaded because of too many incoming requests,
`max_conns` would be a very good temporary measure to protect it.
Because it is so easy to set up.

Progress: https://progress.opensuse.org/issues/132167