Log number of remaining retry attempts for asset uploads
While retries for asset chunks do work correctly, the messages being
logged for failures can be a little confusing. By adding the number of
remaining retries in a way similar to other log messages, it gets a
lot easier to spot what is really going on. And that the "errors" are
actually fairly harmless given the context.

Once this has proven itself in production, we might want to redesign
the event interface a bit and get rid of a few events. There is a bit of
redundancy right now in the logging. But i can see how that might help
find issues in the upload logic as it stands right now.

I've also cleaned up the upload code a little bit to bring it closer to
our current standards, without making changes to the API (yet).

Progress: https://progress.opensuse.org/issues/132167