Remove worker population broadcasts
The `SELECT COUNT(*) FROM workers` query required to broadcast the
worker population is rather expensive, even with an up to date version
of PostgreSQL. And it does not provide all that much value to us. In
fact a hardcoded 1000 would serve us just as well in production.

I've also replaced the very complex status update interval calculation
with a simple min value plus random offset based on maximum value. This
can now also be configured in the worker config or via environment
variables.

The info message sent from the websockets server to the worker as
acknowledgement has not been removed yet, since we do depend on it for
testing quite a bit. And it should have no impact on performance, so
getting rid of it is only a low priority.

Progress: https://progress.opensuse.org/issues/135362