Optimize get_property method
This one surprised me a little bit, since i did not expect the
DBIx::Class overhead to be so significant. At first i wanted to add an
index on the `key` field of the `worker_properties` table, but that made
no measurable difference at all. And i could confirm with
`EXPLAIN ANALYZE` on the production OSD database that the `SELECT` query
used here does not benefit from an additional index.

In the end it was all DBIx::Class overhead that made the difference
getting 1000 `->get_property('JOB_TOKEN')` calls from `1.34s` down to
`0.269s`.

In the greather scheme of things this might not seem like much, but for
the hottest code path in the websocket server it can make a big
difference.

Progress: https://progress.opensuse.org/issues/135362