Make searching for unfinished jobs faster
The `$worker->unfinished_jobs` in the websocket server is the last low
hanging fruit shown by my profiling data. And it is an interesting one
because it has the potential to be a huge improvement, or to be a bit
underwhelming. It is entirely dependent on how many previous jobs the
worker sending the status update has in the database.

Currently the SQL query generated by DBIx::Class ends up using an index
for the `assigned_worker_id` field, but not for `t_finished`. A combined
index for both fields has shown the best imporvements. But Tina brought
up the idea of using the `state` instead of `t_finished`, which is
almost as good.

For the benchmark i've used a real world example from OSD with 1420
previous jobs. Due to DBIx::Class overhead there is a performance floor
around 1.6s, at that point no further optimizations at the PostgreSQL
layer are possible. This patch does not reach the performance floor, but
it has the advantage that no new index is required.

Progress: https://progress.opensuse.org/issues/135362