Log an info message when worker status updates are too frequent
This will give us a hint at when the websocket server is falling behind
with handling worker status updates. It should usually happen when
multiple updates end up in the websocket buffer and are handled right
after another.

In the future we could use this condition to ignore the extra status
updates. Which would give the websocket server a much better chance to
recover on its own, if this was only a temporary performance issue.

Progress: https://progress.opensuse.org/issues/135362