Fix websocket warning not to include the status "working"
This is a small correction to #5303. The worker will send an immediate
"working" status update
[here](https://github.com/os-autoinst/openQA/blob/98e54f9632959aed6a15e3d0afe0a9bdb6249bd4/lib/OpenQA/Worker/Job.pm#L307)
whenever a new job has been started. So that case needs to be
excluded. Since we do expect most problematic cases to involve "free"
status updates, this shouldn't harm the effectiveness of the feature for
detecting scalability issues.

Progress: https://progress.opensuse.org/issues/135362