Fix key navigation on test details
When viewing a test live and test module details are added to the page
dynamically, then the key navigation breaks. It skips steps instead of
just selecting the next/previous step. This is because the event handler
is registered multiple times. This change fixes the issue by making sure
the event handler is only registered once.

Related ticket: https://progress.opensuse.org/issues/113468