Reduce runtime of t/04-scheduler.t
I'm now reusing jobs and also reduced the number of workers. This reduced the
runtime a bit from 2.07 to 1.84.

Considering that the test before only had one worker and two jobs, and that we
want to test code which deals with multiple workers and jobs, I think the
increase to before the original pull request is legitimate.

Issue: https://progress.opensuse.org/issues/129619