Fix dependencies not showing up in graph under certain conditions
When computing the dependencies to show in the dependency graph we prefer
to show the latest child of a parent instead of the child we immediately
found when traversing the list of children. This is implemented as an early
return. In case of this early return we must *not* count the immediate child
as visited. Otherwise we might miss it when traversing the dependency graph
further (and it is the latest child from that perspective).

Related ticket: https://progress.opensuse.org/issues/136154