Add group filter options for /tests
Based on #5388, and reusing most of the same logic, to bring job
group filtering to `/tests` from `/tests/overview`. Similar to the
`?match=...` option, this does not have a UI element yet and can
only be used by manually entering the query parameter. Since the
question of how to extend the filter form with text fields is still
open, i consider it out of scope for this PR.

Progress: https://progress.opensuse.org/issues/134933