Allow searching by comment text on "All tests" page
It works exactly like the `match` parameter and is also implemented as part
of the complex query function.

I tested the performance locally with a dump of the OSD database. Here it
takes around 1.5 seconds to load the finished jobs on the "All tests" page.
When adding `?comment=poo` the time goes up to around 3.4 seconds. When
specifying a comment parameter so that only a few jobs are returned like
`?comment=poo%23119515` the time is even a bit lower (here 2.2 seconds for
this example).
This test means the lookup speed has more than doubled when the `comment`
parameter is used. However, it is still acceptably low for the default
limit of 500 jobs.

Related ticket: https://progress.opensuse.org/issues/151402