Avoid sporadic failures in `t/ui/16-tests_job_next_previous.t`
This test sometimes fails to go to the "Next & previous tab" leading to
failures in subsequent test code. This is likely because the navigation
handling of Bootstrap's JavaScript plugin hasn't been initialized when the
test attempts to click on the tab. This change simply makes the test click
again while waiting for the tab's contents to appear which should fix the
problem.

Related ticket: https://progress.opensuse.org/issues/155068