Ensure all hidden form fields in test overview filter
Issue: https://progress.opensuse.org/issues/154537

The code was skipping elements that already existed in the form, and this
resulted in elements with multiple values to only keep one value.
Now the elements to be added are collected and only added to the form
after the loop.

Bug was introduced in 2f6b7e0fd