Filter settings by result and groupid
This commit provides additional arguments to the job_settings API call in
order to get the ids of the jobs which, in additional the key and the its
value, it will match the job result per job_group. job_group is intended to be
a list for further grouping.

The proposal doesnt demostrate a complete solution.

Obviously the if statement is under further discussion.
An alternative approach could be a separate function, potential named
`blocked` but I am not sure for that particular name as it comes from
internal SUSE concept and it is not assisiated with openQA.

Signed-off-by: ybonatakis <ybonatakis@suse.com>