style: Fix custom footer overflow
On an openQA page using a custom links_footer_left and/or
links_footer_right, e.g.
https://openqa.suse.de/tests/overview?distri=sle&build=20240318-1&result=incomplete
the version string in the footer is cropped despite enough white space
being available. The full version string is readable by scrolling down
but the scroll bar should not be necessary. So no cropping and no scroll
bar should appear if there is enough blank space in both the case of
with or without a custom footer_link content used or not.

This problem was originally introduced with 58747a5ef setting a fixed
height which offsets the version string from the bottom to make it more
readable but does not account for the optional custom_footer inclusion.

This commit fixes this by using a bottom padding instead of prescribing
the overall height of the complete footer box.

Related progress issue: https://progress.opensuse.org/issues/157576