Fix warnings about using undefined values when computing needle details
Apparently the name is not always set. Considering this warning was always
ignored anyway we can also silence it completely. The only bad consequence
is that the redirection won't go to the right step but it looks like this
was never guaranteed to happen anyway.

I also changed the code to use `any` instead of `grep` because that makes
more sense for a check like this.

Related ticket: https://progress.opensuse.org/issues/158808