Avoid race condition when reading step details
The code currently checks whether the file is readable and shows an error
to the user if not. This check is not atomic so we might still run into
e.g. `Can't open file ".../libssh-18.txt": No such file or directory at
OpenQA/Schema/Result/JobModules.pm line 113` leading to a server-side
error and a 500 response. With this change we always show the error to the
user.

This commit also changes another occurrence of the same lookup-pattern to
be consistent (although it is probably not strictly required there).

Related ticket: https://progress.opensuse.org/issues/159447