Improve code for installing assets
Parsing `assetpack.def` via `grep`/`sed` is not the best idea because it
doesn't cover all the assets (as those included via `url(…)` are missing).

This change will invoke the AssetPack plugin itself to get the list of
plugins to be installed which is more reliable as it is definitely in
accordance with AssetPack itself. There's no need to take care of assets
included via `url(…)` separately anymore.