Allow specifying multiple comma-separated values for CORS header
The `Access-Control-Allow-Origin` only allows specifying one concrete
origin or the wildcard `*`. So in order to support multiple allowed origins
we must split the header value server-side and return the header value that
matches the requests origin (if allowed). This is also what is suggested in
the last paragraph of
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin#examples.

This change also only appends the additional value of the Vary header
(which is set in accordance with
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin#cors_and_caching)
to avoid completely overriding it (e.g. preserving the Mojolicious
default).

I tested this locally configuring multiple or just one allowed origins and
verifying via the developer tools that the headers are returned correctly.

Related ticket: https://progress.opensuse.org/issues/159384