Compare SUT package versions in investigation, if available
In addition to showing diff of worker_packages.txt, include also a diff
of sut_packages.txt. Creating the file is up to the test, but do add it
to list of files uploaded at the job end.

Document how to use this feature. And while at it, document also
`worker_packages.txt` and `PACKAGES_CMD` setting.