Enable automatic cloning of Git repos in openqa-bootstrap setups
This enables automatic cloning of Git repositories by the web UI when tests
using Git sources are scheduled. We should not yet remove the automatic
cloning of openSUSE tests and needles in `openqa-boostrap` itself because
we still want to support cloning tests from non-Git sources for now.
Additionally, the download in advance might still be desirable.

I plan to add an according check in openQA-in-openQA tests as part of the
related ticket (https://progress.opensuse.org/issues/161771).