Fix table widths of various admin tables
DataTables sets a fixed width when initializing those tables. This means
that the with doesn't adjust dynamically when the window is resized. This
is not nice and looks especially odd because all the other elements on the
page adjust dynamically.

Adding a width of 100 % prevents DataTables from messing with the width of
these tables. Unfortunately it is not sufficient to add this in CSS so it
needs to be done via a style attribute on every single DataTable.