Fix limiting partial re-scheduling to the relevant test suite
The partial re-scheduling feature relies on specifying the `TEST` parameter
to schedule only a specific test suite. This also works and is tested in
unit tests. However, it only works as long as `scheduled_product_clone_id`
is not used as well. If it is used then the `TEST` parameter as well as all
other parameters are not used at all (as only the parameters of the product
to be cloned are used).

This change simply makes the controller of the "isos post" route consider
the specified parameter also when `scheduled_product_clone_id` is
specified. The newly specified parameters have precedence and possibly
conflicting parameters from the existing scheduled product are removed.

With these changes I got sane behavior when re-scheduling the production
jobs mentioned in the related ticket (locally, using a database dump from
OSD). Before I could also reproduce that basically the whole product was
accidentally re-scheduled.

Related ticket: https://progress.opensuse.org/issues/138593