Fix products being wrongly re-scheduled with the setting `AMP;ASYNC=1`
After f20a0bf597c495941edda7afaa649d657754b7c1 job settings specified as
query parameters are handled also when the `scheduled_product_clone_id`
parameter is present. This revealed the problem that we don't correctly
pass the `async=1` parameter on the products log page. This change fixes
the problem by removing the too excessive escaping.

This also fixes the problem that the async-flag was not actually passed.
This is now tested in the UI test (which would fail without this change).

Related ticket: https://progress.opensuse.org/issues/162608