Be more explicit about the booleans in build_results
There seems to be some ambiguity about exactly what these wind
up being in the hash and then the JSON output generated from that
hash; on Fedora, the tests fail because when the check result is
false, the value seems to be showing up as 0 or '0', not ''.

To try and make this as explicit as possible, let's force the
result of the check to be 0 or 1, and expect the JSON to contain
the same integer. Although in actual fact, integer or string
type doesn't seem to matter to the checks. If we force the
result to be an integer and use strings in the checks - both
json_is and is_deeply - they pass, and ditto if we force the
result to be a string and use integers in the checks. Still,
this seems to be the best we can do.

Signed-off-by: Adam Williamson <awilliam@redhat.com>