Use require for loading OpenQA::Test::Database
It does a

    plan skip_all ...

in its main code, which is effectively an exit.
Since that happens in a BEGIN block currently during a 'use', it won't execute
any INIT blocks any more in perl version >= 5.38. But the -c compile
check is running in an INIT block.

Issue: https://progress.opensuse.org/issues/162866