Prevent failure when restarting Minion-related services
Not all services we are trying to restart might be installed, e.g. in
production setups workers are on separate hosts and therefore no cache
service units are present. This change treats the restart still as success
if not all units can be found.

Related ticket: https://progress.opensuse.org/issues/163013