Avoid potentially slow ws server slowing down live handler as well
After moving the code for dealing with the live view image streaming to the
live handler we should avoid blocking calls in that code because the live
handler is not using multiple processes.

Tested by running a test locally and observing that the `live_log` file in
the pool directory is created/removed as expected. The unit test has also
been extended to cover the error case as well.

Related ticket: https://progress.opensuse.org/issues/163757