Do not retry jobs that were obsoleted
The 'obsoleted' result indicates the job was cancelled because a
new schedule request for the same scenario came in. In this case
it probably doesn't make sense to auto-retry the cancelled job
because the new request indicates the admin wants the job run
with the new settings.

See: https://progress.opensuse.org/issues/164613

Signed-off-by: Adam Williamson <awilliam@redhat.com>