Lower the log level of livestream related problem considering low impact
This error condition has no impact on the job execution. It just means that
the web socket server is not ready yet as we saw in
https://progress.opensuse.org/issues/164709#note-4. The request is already
retried on the client-side so the live view will be able to resume once the
web socket server is ready and the worker connected back. I tested this
locally by entering the "Live" tab while the web socket server is not
running. When starting it the live view updates itself after a few seconds
(without having to switch tabs or reloading the page).