Fix cancelling any pending job from /tests/overview
As @perlpunk found out jobs in "uploading" state can be cancelled from
the tests details page but not the overview page. Also the alignment of
icons looks bad in that case as any row with a job in "uploading" shows
neither the "retrigger" nor the "cancel" action. This commit fixes this
problem for jobs in any "pending state" including not only "uploading"
but also "assigned" or "setup". The change changes the explicit check for
equality to a proper "any" check against the aggregate constant as we
should be using anyway.