Remove git fetch with destination refspec feature
As per git documentation the refspec can formed as <src>:<dst> but dst cannot
use arbitrary expressions (SHA) but an actual ref must be named. Its usage
in the `_git_clone` mistakely adds a SHA ref as destination creating a branch
with its name, which is not supported by git.

As far as now the feature is not used in OpenQA so it can be safely removed.

https://progress.opensuse.org/issues/164937
Signed-off-by: ybonatakis <ybonatakis@suse.com>