Improve git commands in git_clone job
I rearranged and got rid of some calls.

In case of the initial clone, we don't actually need to do a fetch
afterwards, because an initial clone without any restrictions fetches
all refs.
Also, we don't need to retrieve the default remote.

For an update we return early if the remote url doesn't match.
And then only if there is no url fragment with a ref, we fetch
the default remote.

Finally, I removed the second die in case we couldn't get a default remote.
_get_remote_default_branch already dies, the outer die would have
never been reached. This case is now also tested in a new subtest.

(And I improved the regex to avoid having to escape slashes: `\/ \/`)

The full list of commands is now tested, and better readable what commands
are called for which use case.

Related Issue: https://progress.opensuse.org/issues/164886