Require GruTask in database before running a minion job
It can happen that a GruTasks entry is added in a transaction and the minion
job already enqueued and started before the transaction is committed.

In that case we shouldn't start the minion job (the transaction could have
been rollbacked even).

The task needs to be deleted by the minion job in order to go on with
processing a job, and it can't delete it if it doesn't see the entry in the
database.

But requiring the entry before the start of the job is even more correct IMHO.

Issue: https://progress.opensuse.org/issues/165866