Fix linkify to only link to proper URLs
This doesn't check for really well formed URLs, but it limits the matching to
only allowed characters in URIs.
It makes up for escapeForHtml() being already called, so it matches `&amp;`
separately.

Tested with:

    # "> should be excluded
    "https://github.com/os-autoinst/openQA.git"
    <https://github.com/os-autoinst/os-autoinst.git>
    # ) should be included
    https://de.wikipedia.org/wiki/Queen_(Band)
    # query strings should be included
    https://github.com/os-autoinst/openQA?foo=1%20&bar=*2

Single quotes actually can be part of URIs so that's something we should adapt
in our code (and use double quotes, or I remember back in usenet times we used
`<http://...>` whenever referring to a URL in plain text)

Issue: https://progress.opensuse.org/issues/166676