Exclude sub packages with many additional dependencies on SLE
As stated in https://build.suse.de/request/show/343857#comment-4023273 not
all sub packages are installable on SLE. This only concerns the `devel`
and `munin` sub packages which are really optional. So it makes sense to
avoid submitting a huge number of Haskell and Perl packages to SLE and
exclude those packages instead.

Related ticket: https://progress.opensuse.org/issues/165683