Ensure the error responsible for API failures is always added
So far the "last error" was only tracked for REST-API calls (used e.g. for
worker registration and uploads) but not in other cases (e.g. web socket
related errors). This change extends tracking for to cover those cases by
setting the last error to the error messages that is always passed to
`_set_status`. This way they are available to be added to the reason when
declaring the job as incomplete.

Related ticket: https://progress.opensuse.org/issues/163781