Add option to restart jobs upon comment submission
This commit enhances the comment submission functionality with an integrated
job restart option, improving workflow efficiency for users managing multiple
jobs.

For this `restart_job` helper added in `Helpers.pm` and `_restart` in `Job.pm`
had to be refactored to encapsulate the logic and the reusability of the
feature(restart_jobs) in multiple places within the application.

https://progress.opensuse.org/issues/166559

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>