Workaround `scheduler_step` showing up as uncovered
This line shows consistently up as uncovered on Codecov and and in the HTML
report (downloadable via artifacts). The line is however definitely covered
and when executing the tests locally it also appears as such in the HTML
report. The test is also definitely not skipped in the CI. To avoid the
failing checks on all of our PRs I am marking the line as uncoverable for
now.

Related ticket: https://progress.opensuse.org/issues/167272