cli: List 'arch' and 'machine' as supported in API help
Since commit f19726339710 ("Support parameters `arch` and `machine` in job
overview API"), user can filter using those. Add them to the constant shown
to the user in `openqa-cli api -h`. Note that as a side-effect these fields
will now also be explicitly validated as optional non-empty parameters.
This is not expected to have an impact on the API usage.

Signed-off-by: Ricardo B. Marliere <ricardo.marliere@suse.com>