doc: Add back command for starting daemon using unit test setup
Add this comment only at the end as it is not normally recommended to work
on a temporary database possibly causing unwanted changes under `t/data`.
Also add explaining comments so it is clear what this command actually does
and especially that it will cause openQA to no longer take settings from
`database.ini` into account.